Extract ViewState in its own file
Fix circular dependency
