+++ src/screens/actions/useActionsScreen.hook.ts
@@ -4,7 +4,7 @@ import { Action } from '../../core/entities/Action'
 import { GetActionsInteractor } from '../../core/interactor/GetActionsInteractor'
 import { ActionsNavigatorScreenProps } from '../../navigation/ActionsNavigator'
 import { Analytics } from '../../utils/Analytics'
-import { ViewState } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { ActionRowViewModel } from './ActionRowViewModel'
 import { ActionRowViewModelMapper } from './ActionRowViewModelMapper'
