+++ src/screens/doorToDoor/tunnel/brief/DoorToDoorBriefScreen.tsx
@@ -8,7 +8,8 @@ import { StyleSheet, ScrollView, Text, View } from 'react-native'
 import Markdown from 'react-native-markdown-display'
 import { Colors, Spacing, Typography } from '../../../../styles'
 import i18n from '../../../../utils/i18n'
-import { StatefulView, ViewState } from '../../../shared/StatefulView'
+import { StatefulView } from '../../../shared/StatefulView'
+import { ViewState } from '../../../shared/ViewState'
 import { useFocusEffect } from '@react-navigation/core'
 import DoorToDoorRepository from '../../../../data/DoorToDoorRepository'
 import { PrimaryButton } from '../../../shared/Buttons'
