+++ src/screens/doorToDoor/tunnel/opening/TunnelDoorOpeningScreen.tsx
@@ -9,7 +9,8 @@ import { Colors, Spacing, Typography } from '../../../../styles'
 import i18n from '../../../../utils/i18n'
 import { AlertUtils } from '../../../shared/AlertUtils'
 import LoadingOverlay from '../../../shared/LoadingOverlay'
-import { StatefulView, ViewState } from '../../../shared/StatefulView'
+import { StatefulView } from '../../../shared/StatefulView'
+import { ViewState } from '../../../shared/ViewState'
 import { TouchablePlatform } from '../../../shared/TouchablePlatform'
 import { ViewStateUtils } from '../../../shared/ViewStateUtils'
 import { useDoorToDoorTunnelNavigationOptions } from '../DoorToDoorTunnelNavigationHook'
