+++ src/screens/news/useNewsScreen.hook.ts
@@ -5,7 +5,7 @@ import { PaginatedResult } from '../../core/entities/PaginatedResult'
 import NewsRepository from '../../data/NewsRepository'
 import ProfileRepository from '../../data/ProfileRepository'
 import { NewsNavigatorScreenProps } from '../../navigation/NewsNavigator'
-import { ViewState } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { NewsContentViewModel } from './NewsContentViewModel'
 import { NewsContentViewModelMapper } from './NewsContentViewModelMapper'
