+++ src/screens/newsDetail/useNewsDetailScreen.hook.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { News } from '../../core/entities/News'
 import NewsRepository from '../../data/NewsRepository'
 import { ExternalLink } from '../shared/ExternalLink'
-import { ViewState } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { NewsDetailViewModel } from './NewsDetailViewModel'
 import { NewsDetailViewModelMapper } from './NewsDetailViewModelMapper'
