+++ src/screens/personalInformation/centerinterest/CenterOfInterestScreen.tsx
@@ -16,7 +16,8 @@ import { Colors, Spacing, Styles, Typography } from '../../../styles'
 import i18n from '../../../utils/i18n'
 import { PrimaryButton } from '../../shared/Buttons'
 import LoadingOverlay from '../../shared/LoadingOverlay'
-import { StatefulView, ViewState } from '../../shared/StatefulView'
+import { StatefulView } from '../../shared/StatefulView'
+import { ViewState } from '../../shared/ViewState'
 import { CentersOfInterestViewModelMapper } from './CentersOfInterestViewModelMapper'
 import SelectableIconLabelView, {
   SelectableIconLabelViewModel,
