+++ src/screens/personalInformation/notifications/NotificationsScreen.tsx
@@ -18,7 +18,8 @@ import { Colors, Spacing, Typography } from '../../../styles'
 import i18n from '../../../utils/i18n'
 import { AlertUtils } from '../../shared/AlertUtils'
 import LoadingOverlay from '../../shared/LoadingOverlay'
-import { StatefulView, ViewState } from '../../shared/StatefulView'
+import { StatefulView } from '../../shared/StatefulView'
+import { ViewState } from '../../shared/ViewState'
 import { ViewStateUtils } from '../../shared/ViewStateUtils'
 import NotificationRowView from './NotificationRowView'
 import { ID_PUSH, NotificationRowViewModel } from './NotificationViewModel'
