+++ src/screens/phoneCallStatusPicker/PhoneCallStatusPickerScreen.tsx
@@ -24,7 +24,8 @@ import { AlertUtils } from '../shared/AlertUtils'
 import { PrimaryButton } from '../shared/Buttons'
 import LoadingOverlay from '../shared/LoadingOverlay'
 import { VerticalSpacer } from '../shared/Spacer'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { usePreventGoingBack } from '../shared/usePreventGoingBack.hook'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
