+++ src/screens/phoning/PhoningScreen.tsx
@@ -8,7 +8,8 @@ import {
 } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import { Colors, Spacing, Typography } from '../../styles'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { PhoningRowViewModel } from './PhoningRowViewModel'
 import { PhoningViewModel } from './PhoningViewModel'
 import { PhoningViewModelMapper } from './PhoningViewModelMapper'
