+++ src/screens/phoningTutorial/PhoningTutorialScreen.tsx
@@ -10,7 +10,8 @@ import SafeAreaView from 'react-native-safe-area-view'
 import PhoningCampaignRepository from '../../data/PhoningCampaignRepository'
 import { Colors, Spacing, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { useFocusEffect } from '@react-navigation/core'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { ActionsNavigatorScreenProps } from '../../navigation/ActionsNavigator'
