+++ src/screens/pollDetail/PollDetailTools.tsx
@@ -5,7 +5,8 @@ import ToolsRepository from '../../data/ToolsRepository'
 import { Colors, Spacing, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
 import { ExternalLink } from '../shared/ExternalLink'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import PollDetailToolRow from './PollDetailToolRow'
