+++ src/screens/polls/PollsScreen.tsx
@@ -15,7 +15,8 @@ import PollsHeader from './PollsHeader'
 import { Colors } from '../../styles'
 import { PollRowViewModel } from './PollRowViewModel'
 import { PollsScreenViewModelMapper } from './PollsScreenViewModelMapper'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { PollsScreenViewModel } from './PollsScreenViewModel'
 import { GetPollsInteractor } from '../../core/interactor/GetPollsInteractor'
 import { ServerTimeoutError } from '../../core/errors'
