+++ src/screens/profile/ProfileScreen.tsx
@@ -2,7 +2,8 @@ import React, { FC, useCallback, useEffect, useState } from 'react'
 import { StyleSheet, SafeAreaView, Linking } from 'react-native'
 import { useFocusEffect } from '@react-navigation/native'
 import { Colors } from '../../styles'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import ProfileAuthenticated from './ProfileAuthenticated'
 import {
   GetUserProfileInteractor,
