+++ src/screens/regions/RegionScreen.tsx
@@ -6,7 +6,8 @@ import RegionsRepository from '../../data/RegionsRepository'
 import { Colors, Spacing, Styles, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
 import { PrimaryButton } from '../shared/Buttons'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { RegionViewModel } from './RegionViewModel'
 import { RegionViewModelMapper } from './RegionViewModelMapper'
 import { ExternalLink } from '../shared/ExternalLink'
