+++ src/screens/retaliation/useRetaliationDetailScreen.hook.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { Retaliation } from '../../core/entities/Retaliation'
 import RetaliationRepository from '../../data/RetaliationRepository'
 import { RetaliationService } from '../../data/RetaliationService'
-import { ViewState } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { RetaliationDetailViewModel } from './RetaliationDetailViewModel'
 import { RetaliationDetailViewModelMapper } from './RetaliationDetailViewModelMapper'
