+++ src/screens/retaliations/useRetaliationsScreen.hook.ts
@@ -4,7 +4,7 @@ import { Retaliation } from '../../core/entities/Retaliation'
 import RetaliationRepository from '../../data/RetaliationRepository'
 import { RetaliationService } from '../../data/RetaliationService'
 import { ActionsNavigatorScreenProps } from '../../navigation/ActionsNavigator'
-import { ViewState } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ViewStateUtils } from '../shared/ViewStateUtils'
 import { RetaliationListCardViewModel } from './RetaliationListCardViewModel'
 import { RetaliationListCardViewModelMapper } from './RetaliationListCardViewModelMapper'
