+++ src/screens/tools/ToolsScreen.tsx
@@ -4,7 +4,8 @@ import SafeAreaView from 'react-native-safe-area-view'
 import ToolsRepository from '../../data/ToolsRepository'
 import { Colors, Spacing, Typography } from '../../styles'
 import i18n from '../../utils/i18n'
-import { StatefulView, ViewState } from '../shared/StatefulView'
+import { StatefulView } from '../shared/StatefulView'
+import { ViewState } from '../shared/ViewState'
 import { ToolRow } from './ToolRow'
 import { ToolRowViewModel } from './ToolRowViewModel'
 import { ExternalLink } from '../shared/ExternalLink'
